Task authorization policies

+++ app/Http/Controllers/Api/v2/TaskController.php
@@ -8,6 +8,7 @@ use App\Http\Requests\UpdateTaskRequest;
 use App\Http\Resources\TaskResource;
 use App\Models\Task;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Gate;
 
 class TaskController extends Controller
 {
@@ -15,8 +16,10 @@ class TaskController extends Controller
      * Display a listing of the resource.
      * @throws \Throwable
      */
-    public function index()
+    public function index(Request $request)
     {
+        Gate::authorize('viewAny', Task::class);
+
         return request()->user()->tasks()->get()->toResourceCollection();
     }
 
@@ -25,16 +28,20 @@ class TaskController extends Controller
      */
     public function store(StoreTaskRequest $request)
     {
-//        $task = Task::create($request->validated() + ['user_id' => $request->user()->id]);
+        Gate::authorize('create', Task::class);
+
         $task = $request->user()->tasks()->create($request->validated());
+
         return $task->toResource();
     }
 
     /**
      * Display the specified resource.
      */
-    public function show(Task $task)
+    public function show(Task $task, Request $request)
     {
+        Gate::authorize('view', $task);
+
         return $task->toResource();
     }
 
@@ -43,6 +50,8 @@ class TaskController extends Controller
      */
     public function update(UpdateTaskRequest $request, Task $task)
     {
+        Gate::authorize('update', $task);
+
         $task->update($request->validated());
 
         return $task->toResource();
@@ -51,8 +60,10 @@ class TaskController extends Controller
     /**
      * Remove the specified resource from storage.
      */
-    public function destroy(Task $task)
+    public function destroy(Task $task, Request $request)
     {
+        Gate::authorize('delete', $task);
+
         $task->delete();
 
         return response()->noContent();
